docs/requirements.txt
git+https://github.com/sphinx-doc/sphinx
sphinx_rtd_theme
sphinx-autobuild
coloredlogs==15.0.1
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
paho-mqtt==1.6.1
PyYAML==6.0
requests==2.26.0
